02-agents/devops/requirements.txt
python-dotenv
agno[all]
xpander-sdk[agno]
openai
anthropic
mcp